fix: 検索サジェストのキーボード操作対応、IME考慮とテスト強化
- 上下矢印キーでのサジェスト選択切り替えとループ移動を実装
- Enterキーでの選択確定に対応
- IME入力中（e.isComposing）はキーイベントを無視するように修正
- Enterキー押下時に e.preventDefault() を追加し意図しない送信を防止
- サジェストの選択状態管理をstoreに追加
- ロジックをSearchBarに集約し、子コンポーネントへはProps経由でデータを渡すようリファクタリング
- 選択項目のハイライト表示を追加
- SearchBarのテストを更新し、キーボード操作の網羅とカバレッジを確保
- 既存のサジェスト関連テストを新しい設計に合わせて修正
